RAG_app/requirements.txt
qdrant-client==1.12.1

langchain==0.3.12

langchain-community==0.3.12

langchain-core==0.3.25

langchain-huggingface==0.1.2

langchain-openai==0.2.12

langchain-text-splitters==0.3.3

openai==1.57.4

sentence-transformers==3.3.1

python-dotenv==1.0.0

pdfplumber==0.5.28

langchain-qdrant==0.2.0

#qdrant-client==1.5.2

gradio==5.9.1